Move common routines from chtml.ts to common/OutputJax.ts
