+++ mathjax3-ts/output/common/WrapperFactory.ts
@@ -34,7 +34,7 @@ import {MmlNode} from '../../core/MmlTree/MmlNode.js';
  * @template W  The Wrapper type
  * @template C  The WrapperClass type
  */
-export class CommonWrapperFactory<J extends CommonOutputJax<any, any, any, any>,
+export class CommonWrapperFactory<J extends CommonOutputJax<any, any, any, any, any>,
                                   W extends Wrapper<MmlNode, W>,
                                   C extends WrapperClass<MmlNode, W>> extends
 AbstractWrapperFactory<MmlNode, W, C> {
